requirements.txt
Flask
camelot-py